backend/requirements.txt
# Maldonado Repuestos - Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0

# Validation & Serialization
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19

# Payments
mercadopago==2.2.2

# Email
aiosmtplib==3.0.2

# Utilities
python-dotenv==1.0.1
httpx==0.28.1

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
